% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{colExtant}
\alias{colExtant}
\title{Test if a column exists}
\usage{
colExtant(COLNM, SHEET)
}
\arguments{
\item{COLNM}{a column name}
}
\value{
a vector from a column
}
\description{
Test if a column exists
}
