% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{connection_change}
\alias{connection_change}
\title{Opens modal (text-box for input) to ask if save should go ahead}
\usage{
connection_change(INPUT)
}
\value{
a modal object
}
\description{
Opens modal (text-box for input) to ask if save should go ahead
}
