% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{date_as_class}
\alias{date_as_class}
\title{Take a tibble with colnames to match pattern and force to Date if not already}
\usage{
date_as_class(tb, pattern)
}
\arguments{
\item{tb}{tibble}

\item{pattern}{string to match colnames to be forced to Date}
}
\value{
list containing matches and Date values (NB NA coercion likely!)
}
\description{
Take a tibble with colnames to match pattern and force to Date if not already
}
