% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{delete_column}
\alias{delete_column}
\title{Opens modal to ask for columns to delete (only those without data available)}
\usage{
delete_column(VALS_DATA)
}
\arguments{
\item{VALS_DATA}{vals_data object}
}
\value{
a modal object
}
\description{
Opens modal to ask for columns to delete (only those without data available)
}
