% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grep_choice_selectin}
\alias{grep_choice_selectin}
\title{For selectizeInput, grep on set of choices as to what is most likely match,}
\usage{
grep_choice_selectin(choices, label, fail, label_vec = NULL)
}
\arguments{
\item{choices}{vector of choices in selectInput}

\item{label}{string label of selectInput}

\item{fail}{string to return if no matches}
}
\value{
string
}
\description{
For selectizeInput, grep on set of choices as to what is most likely match,
}
