% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grep_choice_selectin_vec}
\alias{grep_choice_selectin_vec}
\title{Set of predetermined named vector for grep_choice_selectin()}
\usage{
grep_choice_selectin_vec()
}
\value{
named vector
}
\description{
Set of predetermined named vector for grep_choice_selectin()
}
