% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{grepSpec}
\alias{grepSpec}
\title{Column grep from vector of terms
grep spectaculare}
\usage{
grepSpec(INVEC, TERMS, INVERTS)
}
\arguments{
\item{INVEC}{a vector to search for match}

\item{TERMS}{a vector of terms to match on}

\item{INVERTS}{a boolean vec of inverted or not greps}
}
\value{
a single character object
}
\description{
Column grep from vector of terms
grep spectaculare
}
