% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{import_cmd_pdf}
\alias{import_cmd_pdf}
\title{Parse CMD PDF reports}
\usage{
import_cmd_pdf(pdf_path, batch = TRUE)
}
\arguments{
\item{pdf_path}{path to PDF}

\item{batch}{logical, if TRUE test all PDFs in pdf_path (default: TRUE)}
}
\value{
named character vector
}
\description{
Parse CMD PDF reports
}
