% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{input_from_xlsx}
\alias{input_from_xlsx}
\title{Input from XLSX}
\usage{
input_from_xlsx(INPUT)
}
\arguments{
\item{INPUT}{object with FILENAMES$datapath}
}
\value{
data.frame of input
}
\description{
Input from XLSX
}
