% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{input_xlsx}
\alias{input_xlsx}
\title{Input XLSX file parsing}
\usage{
input_xlsx(INPUT)
}
\arguments{
\item{INPUT}{object with FILENAMES$datapath}
}
\value{
list of input sheets matching sheet_grep_string()
}
\description{
Input XLSX file parsing
}
