% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{read_sheets_to_list}
\alias{read_sheets_to_list}
\title{Read sheets from XLSX}
\usage{
read_sheets_to_list(FILENAME)
}
\arguments{
\item{FILENAME}{is the XLSX file to read sheets from}
}
\value{
a list object containing Tibble elements
}
\description{
Read sheets from XLSX
}
