% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_input}
\alias{parse_input}
\title{Tests for previous input data, and/or takes input from user
ensures that data is formatted correctly, saved correctly}
\usage{
parse_input(INPUT, VALS_DATA)
}
\value{
a Tibble object
}
\description{
Tests for previous input data, and/or takes input from user
ensures that data is formatted correctly, saved correctly
}
