% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{macroDissect}
\alias{macroDissect}
\title{Parse macrodissection into correct format
ensures that data is formatted correctly}
\usage{
macroDissect(INPUT)
}
\value{
a Tibble object
}
\description{
Parse macrodissection into correct format
ensures that data is formatted correctly
}
