% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_add_column}
\alias{mod_add_column}
\title{Opens modal (text-box for input) to ask for name of new column}
\usage{
mod_add_column(INPUT)
}
\value{
a modal object
}
\description{
Opens modal (text-box for input) to ask for name of new column
}
