% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_load_data_proceed}
\alias{mod_load_data_proceed}
\title{Tell user a table they are trying to create exists and will be appended with new data}
\usage{
mod_load_data_proceed(INPUT)
}
\value{
a modal object
}
\description{
Tell user a table they are trying to create exists and will be appended with new data
}
