% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_map_columns}
\alias{mod_map_columns}
\title{Opens modal to ask for map from data column to table column}
\usage{
mod_map_columns(INPUT, VALS_TIB, NAME = NULL)
}
\arguments{
\item{VALS_TIB}{input tibble of new data to tabulate}

\item{SHEETNAME}{for XLSX input allow sheet name to be used as a column}
}
\value{
a modal object
}
\description{
Opens modal to ask for map from data column to table column
}
