% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_rename_column}
\alias{mod_rename_column}
\title{Opens modal (dropdowns for input) to ask for column to rename}
\usage{
mod_rename_column(VALS_DATA)
}
\arguments{
\item{VALS_DATA}{named list of colnames of data}
}
\value{
a modal object
}
\description{
Opens modal (dropdowns for input) to ask for column to rename
}
