% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_sure_to_save}
\alias{mod_sure_to_save}
\title{Opens modal (text-box for input) to ask if save should go ahead}
\usage{
mod_sure_to_save(INPUT)
}
\arguments{
\item{INPUT}{object}
}
\value{
a modal object
}
\description{
Opens modal (text-box for input) to ask if save should go ahead
}
