% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_test_db_con}
\alias{mod_test_db_con}
\title{Test if connection to Postgresql can be made}
\usage{
mod_test_db_con(INPUT)
}
\value{
a modal object
}
\description{
Test if connection to Postgresql can be made
}
