% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_test_db_results}
\alias{mod_test_db_results}
\title{Test connection to Postgresql, handle success and error}
\usage{
mod_test_db_results(INPUT, CON)
}
\value{
a modal object
}
\description{
Test connection to Postgresql, handle success and error
}
