% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{mod_validate_user_cond}
\alias{mod_validate_user_cond}
\title{Conditionally show advanced connection options for validate_user_nt
from: https://stackoverflow.com/questions/65168516}
\usage{
mod_validate_user_cond(INPUT)
}
\arguments{
\item{INPUT}{input reactive values}
}
\description{
Conditionally show advanced connection options for validate_user_nt
from: https://stackoverflow.com/questions/65168516
}
