% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{mutationStatus}
\alias{mutationStatus}
\title{Parse mutation status into correct format
ensures that data is formatted correctly}
\usage{
mutationStatus(INPUT)
}
\value{
a Tibble object
}
\description{
Parse mutation status into correct format
ensures that data is formatted correctly
}
