% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{new_table_reqs}
\alias{new_table_reqs}
\title{Required columns for new table}
\usage{
new_table_reqs()
}
\value{
vector of required values for table, if all NA or "-" row is removed
}
\description{
Required columns for new table
}
