% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_arr_col}
\alias{obsev_arr_col}
\title{Method to reorder columns in table}
\usage{
obsev_arr_col(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Method to reorder columns in table
}
