% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_col_del}
\alias{obsev_col_del}
\title{Method to delete column into table}
\usage{
obsev_col_del(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Method to delete column into table
}
