% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_db_conxn}
\alias{obsev_db_conxn}
\title{Allow change of connection credentials}
\usage{
obsev_db_conxn(INPUT, OUTPUT)
}
\arguments{
\item{INPUT}{session input}

\item{OUTPUT}{session OUTPUT}
}
\description{
Allow change of connection credentials
}
