% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_del_col}
\alias{obsev_del_col}
\title{Delete column into table}
\usage{
obsev_del_col(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Delete column into table
}
