% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_disconnex}
\alias{obsev_disconnex}
\title{Allow disconnection (useful if too many previous connections and for logoff)}
\usage{
obsev_disconnex(INPUT)
}
\arguments{
\item{INPUT}{session input}
}
\description{
Allow disconnection (useful if too many previous connections and for logoff)
}
