% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_extantabled}
\alias{obsev_extantabled}
\title{Allow change of connection credentials}
\usage{
obsev_extantabled(INPUT, CON, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{CON}{connection reactiveVal}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Allow change of connection credentials
}
