% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_FILENAMES}
\alias{obsev_FILENAMES}
\title{Take filename INPUT and parse}
\usage{
obsev_FILENAMES(INPUT, CON, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{CON}{connection reactiveVal}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Take filename INPUT and parse
}
