% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_go_loaddata}
\alias{obsev_go_loaddata}
\title{Load data into current table}
\usage{
obsev_go_loaddata(INPUT, CON, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{CON}{connection reactiveVal}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Load data into current table
}
