% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_go_save}
\alias{obsev_go_save}
\title{Save table to db}
\usage{
obsev_go_save(INPUT, CON, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{CON}{connection reactiveVal}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Save table to db
}
