% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_ins_col}
\alias{obsev_ins_col}
\title{Method to add column into table}
\usage{
obsev_ins_col(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Method to add column into table
}
