% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_ord_col}
\alias{obsev_ord_col}
\title{Method to initiate reorder columns in table}
\usage{
obsev_ord_col(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Method to initiate reorder columns in table
}
