% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_ren_col}
\alias{obsev_ren_col}
\title{Method to rename columns in table}
\usage{
obsev_ren_col(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Method to rename columns in table
}
