% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_rename_col}
\alias{obsev_rename_col}
\title{Reordering columns in table}
\usage{
obsev_rename_col(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Reordering columns in table
}
