% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_uni_col}
\alias{obsev_uni_col}
\title{Uniquify columns in table}
\usage{
obsev_uni_col(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Uniquify columns in table
}
