% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_uniq_col}
\alias{obsev_uniq_col}
\title{Column data uniquified and output}
\usage{
obsev_uniq_col(INPUT, VALS_DATA, OUTPUT)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}

\item{OUTPUT}{session OUTPUT}

\item{CON}{database connection}
}
\description{
Column data uniquified and output
}
