% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_userpass}
\alias{obsev_userpass}
\title{Test connection credentials and connect, specify if table is extant}
\usage{
obsev_userpass(INPUT, CON)
}
\arguments{
\item{INPUT}{session input}

\item{CON}{connection reactiveVal}
}
\description{
Test connection credentials and connect, specify if table is extant
}
