% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_valsdata_new}
\alias{obsev_valsdata_new}
\title{Parse Dates and other Data from VALS_DATA$New and binds to VALS_DATA$Data}
\usage{
obsev_valsdata_new(VALS_DATA)
}
\arguments{
\item{VALS_DATA}{reactive with $Data element of data (can be empty) and $New (nonempty)}
}
\value{
none, changes VALS_DATA$Data to df with VALS_DATA$New added
}
\description{
Parse Dates and other Data from VALS_DATA$New and binds to VALS_DATA$Data
}
