% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{otherSummary}
\alias{otherSummary}
\title{Allow choice to return 'other' as a value from summary tables}
\usage{
otherSummary(COLNM, CHOICE)
}
\value{
a Tibble object
}
\description{
Allow choice to return 'other' as a value from summary tables
}
