% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parseGeneral}
\alias{parseGeneral}
\title{General parser for data structures incoming
ensures that data is formatted correctly}
\usage{
parseGeneral(SHEET, NAME)
}
\value{
a Tibble object
}
\description{
General parser for data structures incoming
ensures that data is formatted correctly
}
