% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{refHospital}
\alias{refHospital}
\title{Parse referring hospital into correct format
ensures that data is formatted correctly}
\usage{
refHospital(INPUT)
}
\value{
a Tibble object
}
\description{
Parse referring hospital into correct format
ensures that data is formatted correctly
}
