% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{renameParse}
\alias{renameParse}
\title{Rename parsed data into output format
ensures that data is formatted correctly}
\usage{
renameParse(INPUT)
}
\value{
a Tibble object
}
\description{
Rename parsed data into output format
ensures that data is formatted correctly
}
