% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_observe.R
\name{render_simple_maintable}
\alias{render_simple_maintable}
\title{DT:renderDataTable filtering
The resulting output should be caught by output$maintable}
\usage{
render_simple_maintable(INPUT, DATAER)
}
\value{
a DataTable object
}
\description{
DT:renderDataTable filtering
The resulting output should be caught by output$maintable
}
