% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_observe.R
\name{render_unique_maintable}
\alias{render_unique_maintable}
\title{DT:renderDataTable}
\usage{
render_unique_maintable(VALS_DATA)
}
\arguments{
\item{VALS_DATA}{data frame with columns to display uniquely}
}
\value{
a data.frame object
}
\description{
DT:renderDataTable
}
