% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{siteCode}
\alias{siteCode}
\title{Parse site code
ensures that data is formatted correctly}
\usage{
siteCode(INPUT)
}
\value{
a Tibble object
}
\description{
Parse site code
ensures that data is formatted correctly
}
