% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{siteSource}
\alias{siteSource}
\title{Parse site source into correct format
ensures that data is formatted correctly}
\usage{
siteSource(INPUT)
}
\value{
a Tibble object
}
\description{
Parse site source into correct format
ensures that data is formatted correctly
}
