% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tb_col_delim_split}
\alias{tb_col_delim_split}
\title{Take a tibble column and split on delimiter returning index x of vector from split}
\usage{
tb_col_delim_split(tb, col, delim, x)
}
\arguments{
\item{tb}{tibble with column 'col'}

\item{col}{colname to split on delim and take xth entry}

\item{delim}{delimiter to split entries in tb\link{,col}}

\item{x}{index of element to return}
}
\value{
string
}
\description{
Take a tibble column and split on delimiter returning index x of vector from split
}
