% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{test_coding}
\alias{test_coding}
\title{Make Test_Code from sheet NAME and other Test_Codes}
\usage{
test_coding(CODE, NAME)
}
\arguments{
\item{CODE}{column of codes extant in sheet ('test_code')}

\item{NAME}{of sheet (takes first element split on whitespace)}
}
\value{
vector of values
}
\description{
Make Test_Code from sheet NAME and other Test_Codes
}
