% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{test_db_con}
\alias{test_db_con}
\title{Test connection to Postgresql, handle success and error}
\usage{
test_db_con(INPUT)
}
\value{
a modal object
}
\description{
Test connection to Postgresql, handle success and error
}
