% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{tissueCode}
\alias{tissueCode}
\title{Parse tissue code
ensures that data is formatted correctly}
\usage{
tissueCode(INPUT)
}
\value{
a Tibble object
}
\description{
Parse tissue code
ensures that data is formatted correctly
}
