% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{tissueSource}
\alias{tissueSource}
\title{Parse tissue source into correct format
ensures that data is formatted correctly}
\usage{
tissueSource(INPUT)
}
\value{
a Tibble object
}
\description{
Parse tissue source into correct format
ensures that data is formatted correctly
}
