% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{uniq_columns}
\alias{uniq_columns}
\title{Opens modal (dropdowns for input) to ask for columns to display uniqly in table}
\usage{
uniq_columns(VALS_DATA)
}
\arguments{
\item{VALS_DATA}{named list of colnames of data}
}
\value{
a modal object
}
\description{
Opens modal (dropdowns for input) to ask for columns to display uniqly in table
}
