% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{validate_user_nt}
\alias{validate_user_nt}
\title{Opens modal (text-box for input) to ask for user credentials}
\usage{
validate_user_nt(INPUT)
}
\value{
a modal object
}
\description{
Opens modal (text-box for input) to ask for user credentials
}
