% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{year_shootout}
\alias{year_shootout}
\title{Make Year based on one of the 4 dates included}
\usage{
year_shootout(Y1, Y2, Y3, Y4, NAME)
}
\arguments{
\item{Y1}{date the first}

\item{Y2}{date the second}

\item{Y3}{date the 3rd}

\item{Y4}{date the dth}
}
\value{
single most represented year
}
\description{
Make Year based on one of the 4 dates included
}
