% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{import_cmd_pdfs}
\alias{import_cmd_pdfs}
\title{Parse CMD PDF reports, very hacky}
\usage{
import_cmd_pdfs(pdf_path)
}
\arguments{
\item{pdf_path}{path to PDF}
}
\value{
named character vector
}
\description{
Parse CMD PDF reports, very hacky
}
