% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{newtable_exists}
\alias{newtable_exists}
\title{Tell user a table they are trying to create exists and will be appended with new data}
\usage{
newtable_exists(INPUT)
}
\value{
a modal object
}
\description{
Tell user a table they are trying to create exists and will be appended with new data
}
