% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_go_askdata}
\alias{obsev_go_askdata}
\title{Data read}
\usage{
obsev_go_askdata(INPUT, CON, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{CON}{connection reactiveVal}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Data read
}
