% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_go_rds}
\alias{obsev_go_rds}
\title{Go saving RDS}
\usage{
obsev_go_rds(INPUT, VALS_DATA)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}
}
\description{
Go saving RDS
}
