% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_events.R
\name{obsev_show_tab}
\alias{obsev_show_tab}
\title{Show table}
\usage{
obsev_show_tab(INPUT, VALS_DATA, OUTPUT)
}
\arguments{
\item{INPUT}{session input}

\item{VALS_DATA}{data reactiveVal}

\item{OUTPUT}{session OUTPUT}
}
\description{
Show table
}
